Streamlined language additions + running files over REPLs
I built more onto my previous solution, making it easier to be able to run both compiled and interpreted languages without having to keep adding to a switch case for each language being added. I'm working on trying to use onExit() over onData() and (hopefully) be able to aggregate the data without duplicating it due to how weird the output chunking from onData is. Also every language now stores all the inputted code into a file, and that file gets run and the output gets displayed instead of worrying about messing with each language's REPL implementations. I thought adding this would fix the error with the Javascript file not running the first line (from what I can tell) when there are multiple lines in the file, but it did not.

TODO:
- Add more languages (C/C++, Java / Kotlin, Zig(?))
- Fix the problem with the JS runner
- Aggregate the output and (cleanly) display it
- Begin working on the UI