Error message formatting + aggregation of output
Working with IPty has been a bit frustrating, but I was able to semi-successfully format and process the data recieved after running the commands for both interpreted and compiled languages (although Rust is the only language that it was tested with). Also, aggregating the data received from onData and emitting on the socket after onExit is the method I decided to use. Doing so makes it easier to "guarantee" the formatting of output being received, cleaned, and displayed properly. It also fixed the issue with JS not running the first line in the file. I also changed the editor's theme to look more familiar :)

TODO:
- Add more languages (C/C++, Java/Kotlin, Zig(?))
- Begin working on the UI for the complete web app.