More languages!!
I streamlined the process for adding new languages even more than before. Instead of splitting each language by whether it's "compiled" or not (you get into grey areas, such as 'Does *transpiling* count as compilation? What about JVM languages??'), it spawns a pty zsh instance and runs all languages from there. Each language has information attached to it as to how to run it's file type (and compile if necessary) and exit cleanly, as then onExit is called and the aggregated data is shown. Some special cases have to be handled (e.g. Go "warning" messages), but it seems to be working for the most part. The languages that were added are TypeScript, Java, Kotlin, Elixir, and Go. More languages are planned to be added in the future, but for now there are more important priorities to take care of.

TODO:
- Begin working on UI
- Aggregate Go's warning messages into output.
- Add core features such as user login, game-mode selection, etc.