middleware fix + design changes in ide + theme additions + logout sidebar
Fixed an issue within the middleware where the middleware would only activate for two paths (/ and /login). Whenever a new path is added as a route, it must be added to the list of routes within middleware.ts. Added logout functionality to the SideBar (and removed the horrid logout option from the ide). Also changed up formatting for ide and changed theme to a more dark-grey theme over a blue one. In the future, there will be a switch to toggle between light and dark mode on the website.

TODO:
- Work on functionality for one of the three game-modes (Solo first?)
- Setup user information being displayed in SideBar when logged in