fixed db insertions + favicon + basic layout for solo arena mode
There was an issue with insertions not being put into the table, and after testing it was found that it was due to there being duplicate ID values attempting to be inserted into the table when ID's have to be unique, so no insertions (and no errors) were occurring. Added a favicon logo (shurikan) for the website, as well as begun the layout for the Solo page and made minor fixes / changes around the project.

TODO:
- Finish the basic layout for Solo
- Finish the layout for the Question cards
- Create Questions table and retrieve that data and display on Solo properly
- Functionality for filtering questions by language + difficulty
- Begin level / ranking system for users
- Display logged in user's personal data on sidebar
- Research / plan for solving questions functionality