Dynamic page route for individual questions + filtering and search functionality
Implemented the search filtering for Solo questions, as well as filtering by language and difficulty. However, the filtering by language should probably be changed to a multi-select of sorts, allowing for people to select multiple languages to search for. Also allowed for multiple languages per question (hence multiple icons per question), and created a dynamic empty page for each question, which will be implemented soon.

TODO:
- Work on layout for individual questions
- Move temporary DB impl to supabase
- Work on level / ranking system for users
- Display logged in user's profile information on sidebar (or profile page?)
- Continue research on implementing testing / submission features for each question.