Data cleaning + basic question layout
Began using stripAnsi package in order to make cleaning data from terminal easier. While it removes the color highlighting, which will be annoying to re-implement manually in the future, it allows for panic!() messages to pass through, which is a worthy tradeoff, since assert_eq!() causes panics (which is what is being used to test user input to solution input). Also made changes to the fake db to include testing, starter, and solution code.