Rust question solution functionality
Success! Not must testing was done, but it seems that running Rust problems works properly. Since there's no interface / api for users to be able to create their own questions / tests / solutions, they have to be added manually (future implementation?). Unfortunately, it has broken the ability to run rust through the Online IDE, but that should be changed in the future (separate servers? or tags to differentiate server tasks?).

TODO:
- Work on layout for question completion page.
- Move temporary DB impl to supabase.
- Work on level / ranking system for users.
- Display logged in user's profile information on sidebar (or profile page?).
- Continue research on implementing testing / submission features for different languages.