Sidebar data persists through pageloads + public/uploads
Dealing with the Sidebar constantly making requests to supabase asyncronously and attempting to display that data was frustrating, since you could see the data loading in whenever the page was refreshed / switched to a different route. Now, user data is stored within localStorage on sign-in, and can be accessed from any route. However, this means that when completing a question, earning points, or changing data via settings, the data does not get updated until the user logs out and re-logs back in (which will be added in the next commit).